Move config to be public facing
