+++ vec_inf/client/__init__.py
@@ -5,8 +5,8 @@ using `vec_inf`. It is an alternative to the command-line interface, and allows
 users direct control over the lifecycle of inference servers via python scripts.
 """
 
-from vec_inf.client._config import ModelConfig
 from vec_inf.client.api import VecInfClient
+from vec_inf.client.config import ModelConfig
 from vec_inf.client.models import (
     LaunchOptions,
     LaunchOptionsDict,
