+++ vec_inf/client/_utils.py
@@ -15,7 +15,7 @@ import requests
 import yaml
 
 from vec_inf.client._client_vars import MODEL_READY_SIGNATURE
-from vec_inf.client._config import ModelConfig
+from vec_inf.client.config import ModelConfig
 from vec_inf.client.models import ModelStatus
 from vec_inf.client.slurm_vars import CACHED_CONFIG
